Refactored access management config parser tests
